refactor: rename user-related files
