+++ package.json
@@ -6,8 +6,8 @@
   "main": "src/main.js",
   "scripts": {
     "test": "jest",
-    "start": "node ./src/main.js",
-    "dev": "node ./src/main.js",
+    "start": "node ./src/app.js",
+    "dev": "node ./src/app.js",
     "build": "prisma generate && npm run seed",
     "postinstall": "prisma generate",
     "seed": "node ./prisma/seed.js"
